Reported damage amount now can have up to 2 decimal places (i.e. "1.23") instead of just whole/integer numbers

+++ src/com/wimbli/serverevents/DeathThread.java
@@ -38,7 +38,7 @@ public class DeathThread implements Runnable {
     private Player player;
 	private EntityDamageEvent.DamageCause damageCause;
 	private Entity damager;
-	private int damageAmount;
+	private double damageAmount;
     
     protected static final Logger log = Logger.getLogger("Minecraft");
